requirements.txt
sqlalchemy
os
fastapi
python-dotenv
openai
psycopg2-binary
setuptools
streamlit